split libraries in preparation for dynamic imports
