+++ sites/fast-creator/app/configs/index.ts
@@ -1,1 +1,10 @@
-export * from "./linked-data-examples";
+import { fastLinkedDataExamples } from "./library.fast.examples";
+import { nativeLinkedDataExamples } from "./library.native.examples";
+import { ExampleData } from "./typings";
+
+const linkedDataExamples: ExampleData = {
+    ...fastLinkedDataExamples,
+    ...nativeLinkedDataExamples,
+};
+
+export { linkedDataExamples };
