+++ sites/fast-creator/app/configs/library.fast.examples.ts
@@ -1,0 +1,75 @@
+import { fastComponentSchemas } from "@microsoft/site-utilities";
+import { ExampleData } from "./typings";
+import {
+    fastAnchorTag,
+    fastBadgeTag,
+    fastButtonTag,
+    fastCardTag,
+    fastCheckboxTag,
+    fastDialogTag,
+    fastDividerTag,
+    fastFlipperTag,
+    fastMenuItemTag,
+    fastMenuTag,
+    fastProgressRingTag,
+    fastProgressTag,
+    fastRadioGroupTag,
+    fastRadioTag,
+    fastSliderLabelTag,
+    fastSliderTag,
+    fastSwitchTag,
+    fastTabPanelTag,
+    fastTabsTag,
+    fastTabTag,
+    fastTextAreaTag,
+    fastTextFieldTag,
+} from "./library.fast.tags";
+import {
+    fastAnchorExample,
+    fastBadgeExample,
+    fastButtonExample,
+    fastCardExample,
+    fastCheckboxExample,
+    fastDialogExample,
+    fastDividerExample,
+    fastFlipperExample,
+    fastMenuExample,
+    fastMenuItemExample,
+    fastProgressExample,
+    fastProgressRingExample,
+    fastRadioExample,
+    fastRadioGroupExample,
+    fastSliderExample,
+    fastSliderLabelExample,
+    fastSwitchExample,
+    fastTabExample,
+    fastTabPanelExample,
+    fastTabsExample,
+    fastTextAreaExample,
+    fastTextFieldExample,
+} from "./library.fast";
+
+export const fastLinkedDataExamples: ExampleData = {
+    [fastComponentSchemas[fastAnchorTag].id]: [fastAnchorExample],
+    [fastComponentSchemas[fastButtonTag].id]: [fastButtonExample],
+    [fastComponentSchemas[fastBadgeTag].id]: [fastBadgeExample],
+    [fastComponentSchemas[fastCardTag].id]: [fastCardExample],
+    [fastComponentSchemas[fastCheckboxTag].id]: [fastCheckboxExample],
+    [fastComponentSchemas[fastDialogTag].id]: [fastDialogExample],
+    [fastComponentSchemas[fastDividerTag].id]: [fastDividerExample],
+    [fastComponentSchemas[fastFlipperTag].id]: [fastFlipperExample],
+    [fastComponentSchemas[fastMenuTag].id]: [fastMenuExample],
+    [fastComponentSchemas[fastMenuItemTag].id]: [fastMenuItemExample],
+    [fastComponentSchemas[fastProgressTag].id]: [fastProgressExample],
+    [fastComponentSchemas[fastProgressRingTag].id]: [fastProgressRingExample],
+    [fastComponentSchemas[fastRadioGroupTag].id]: [fastRadioGroupExample],
+    [fastComponentSchemas[fastRadioTag].id]: [fastRadioExample],
+    [fastComponentSchemas[fastSliderTag].id]: [fastSliderExample],
+    [fastComponentSchemas[fastSliderLabelTag].id]: [fastSliderLabelExample],
+    [fastComponentSchemas[fastSwitchTag].id]: [fastSwitchExample],
+    [fastComponentSchemas[fastTabsTag].id]: [fastTabsExample],
+    [fastComponentSchemas[fastTabTag].id]: [fastTabExample],
+    [fastComponentSchemas[fastTabPanelTag].id]: [fastTabPanelExample],
+    [fastComponentSchemas[fastTextAreaTag].id]: [fastTextAreaExample],
+    [fastComponentSchemas[fastTextFieldTag].id]: [fastTextFieldExample],
+};
