+++ sites/fast-creator/app/configs/library.fast.tags.ts
@@ -1,0 +1,25 @@
+/**
+ * FAST component tags
+ */
+export const fastAnchorTag = "fast-anchor";
+export const fastBadgeTag = "fast-badge";
+export const fastButtonTag = "fast-button";
+export const fastCardTag = "fast-card";
+export const fastCheckboxTag = "fast-checkbox";
+export const fastDialogTag = "fast-dialog";
+export const fastDividerTag = "fast-divider";
+export const fastFlipperTag = "fast-flipper";
+export const fastMenuTag = "fast-menu";
+export const fastProgressTag = "fast-progress";
+export const fastProgressRingTag = "fast-progress-ring";
+export const fastRadioGroupTag = "fast-radio-group";
+export const fastRadioTag = "fast-radio";
+export const fastSliderTag = "fast-slider";
+export const fastSliderLabelTag = "fast-slider-label";
+export const fastSwitchTag = "fast-switch";
+export const fastTabsTag = "fast-tabs";
+export const fastTabTag = "fast-tab";
+export const fastTabPanelTag = "fast-tab-panel";
+export const fastMenuItemTag = "fast-menu-item";
+export const fastTextAreaTag = "fast-text-area";
+export const fastTextFieldTag = "fast-text-field";
