+++ sites/fast-creator/app/configs/library.native.tags.ts
@@ -1,0 +1,14 @@
+/**
+ * Native element tags
+ */
+export const divTag = "div";
+export const imageTag = "img";
+export const paragraphTag = "p";
+export const spanTag = "span";
+export const heading1Tag = "h1";
+export const heading2Tag = "h2";
+export const heading3Tag = "h3";
+export const heading4Tag = "h4";
+export const heading5Tag = "h5";
+export const heading6Tag = "h6";
+export const labelTag = "label";
