+++ sites/fast-creator/app/creator.tsx
@@ -41,7 +41,8 @@ import { neutralLayerL1, StandardLuminance } from "@microsoft/fast-components";
 import * as monaco from "monaco-editor/esm/vs/editor/editor.api";
 import { monacoAdapterId } from "@microsoft/fast-tooling/dist/esm/message-system-service/monaco-adapter.service";
 import { CreatorState, FormId, ProjectFile } from "./creator.props";
-import { divTag, linkedDataExamples } from "./configs";
+import { linkedDataExamples } from "./configs";
+import { divTag } from "./configs/library.native.tags";
 import { ProjectFileTransfer } from "./components";
 import { previewReady } from "./preview";
 import { Footer } from "./site-footer";
